backend/requirements.txt
# DocFlow HR Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.26.0

# Forms and File Upload
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development (optional)
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.8.0
# ruff>=0.1.0
